Deduplicate Boost/GTest calculator examples
Considering that the only difference is a single comparison macro: just
define that macro to be the same as GTest when building for Boost.Test.
