+++ examples/Calc/features/step_definitions/CalculatorSteps.cpp
@@ -1,0 +1,36 @@
+#include <cucumber-cpp/autodetect.hpp>
+
+#include <Calculator.h>
+
+using cucumber::ScenarioScope;
+
+struct CalcCtx {
+    Calculator calc;
+    double result;
+};
+
+GIVEN("^I have entered (\\d+) into the calculator$") {
+    REGEX_PARAM(double, n);
+    ScenarioScope<CalcCtx> context;
+
+    context->calc.push(n);
+}
+
+WHEN("^I press add") {
+    ScenarioScope<CalcCtx> context;
+
+    context->result = context->calc.add();
+}
+
+WHEN("^I press divide") {
+    ScenarioScope<CalcCtx> context;
+
+    context->result = context->calc.divide();
+}
+
+THEN("^the result should be (.*) on the screen$") {
+    REGEX_PARAM(double, expected);
+    ScenarioScope<CalcCtx> context;
+
+    EXPECT_EQ(expected, context->result);
+}
