+++ examples/Calc/features/step_definitions/GTestCalculatorSteps.cpp
@@ -1,38 +1,2 @@
 #include <gtest/gtest.h>
-#include <cucumber-cpp/autodetect.hpp>
-
-#include <Calculator.h>
-
-using cucumber::ScenarioScope;
-
-struct CalcCtx {
-    Calculator calc;
-    double result;
-};
-
-GIVEN("^I have entered (\\d+) into the calculator$") {
-    REGEX_PARAM(double, n);
-    ScenarioScope<CalcCtx> context;
-
-    context->calc.push(n);
-}
-
-WHEN("^I press add") {
-    ScenarioScope<CalcCtx> context;
-
-    context->result = context->calc.add();
-}
-
-WHEN("^I press divide") {
-    ScenarioScope<CalcCtx> context;
-
-    context->result = context->calc.divide();
-}
-
-THEN("^the result should be (.*) on the screen$") {
-    REGEX_PARAM(double, expected);
-    ScenarioScope<CalcCtx> context;
-
-    EXPECT_EQ(expected, context->result);
-}
-
+#include "CalculatorSteps.cpp"
